FAQ - validation - login required and staff or superuser - added

+++ faq/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, get_object_or_404, redirect, reverse
 from .models import Faq
 from .forms import FaqForm
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
@@ -15,8 +16,13 @@ def all_faq(request):
     return render(request, 'faq/faq.html', context)
 
 
+@login_required
 def add_question(request):
     """ Add question into FAQ """
+    if not request.user.is_staff or not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = FaqForm(request.POST, request.FILES)
         if form.is_valid():
@@ -36,8 +42,13 @@ def add_question(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_question(request, faq_id):
     """ Edit question in FAQ """
+    if not request.user.is_staff or not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     question = get_object_or_404(Faq, pk=faq_id)
     if request.method == 'POST':
         form = FaqForm(request.POST, request.FILES, instance=question)
@@ -59,8 +70,12 @@ def edit_question(request, faq_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_question(request, faq_id):
     """ Delete a question in FAQ """
+    if not request.user.is_staff or not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
 
     question = get_object_or_404(Faq, pk=faq_id)
     question.delete()
